Moved querying logic out of the Live page.
