+++ lib/slax_web/live/chat_room/page/room_page.ex
@@ -1,14 +1,7 @@
 defmodule SlaxWeb.Live.ChatRoom.Page do
   use SlaxWeb, :live_view
 
-  import Ecto.Query
-
-  alias SlaxWeb.Live.ChatRoom.Helpers.Room
-  alias Slax.Repo
-  alias Slax.Chat.Room
-
-  @paginated_results_limit 100
-  @max_number_of_rooms 10_000_000
+  alias SlaxWeb.Live.ChatRoom.{Querying.RoomQueries, Helpers.Room}
 
   # When a LiveView page is opened, the first thing that gets done is mounting the page. It's the LiveView entrypoint.
   # mount/3 is called 2x; once of the page loads and then again to establish the live websocket.
@@ -17,34 +10,11 @@ defmodule SlaxWeb.Live.ChatRoom.Page do
   # Source: https://hexdocs.pm/phoenix_live_view/Phoenix.LiveView.html#c:mount/3
 
   def mount(_params, _session, socket) do
-    slax_chat_room =
-      build_room_query()
-      |> fetch_first_room()
+    slax_chat_room = RoomQueries.get_first_room()
 
-    socket =
-      socket
-      |> assign(slax_room: slax_chat_room)
+    socket = assign(socket, slax_room: slax_chat_room)
 
     {:ok, socket}
-  end
-
-  defp build_room_query(opts \\ []) do
-    Room
-    |> select([r], r)
-    |> apply_limit(opts)
-    |> apply_offset(opts)
-  end
-
-  defp fetch_first_room(query), do: Repo.one(query)
-
-  defp apply_limit(query, opts) do
-    limit = Keyword.get(opts, :limit, @paginated_results_limit)
-    from(r in query, limit: ^limit)
-  end
-
-  defp apply_offset(query, opts) do
-    offset = Keyword.get(opts, :offset, 0)
-    from(r in query, offset: ^offset)
   end
 
   def render(assigns) do
